requirements.txt
future==0.16.0
oauthlib==2.0.1
python-twitter==3.2
requests==2.12.4
requests-oauthlib==0.7.0
